# planner: keep original select fields names for non-prep plan cache