--- a/parser/format/format.go
+++ b/parser/format/format.go
@@ -238,6 +238,7 @@ const (
 	RestoreWithTTLEnableOff
 	RestoreWithoutSchemaName
 	RestoreWithoutTableName
+	RestoreForNonPrepPlanCache
 )
 
 const (
@@ -337,6 +338,12 @@ func (rf RestoreFlags) HasSkipPlacementRuleForRestoreFlag() bool {
 // HasRestoreWithTTLEnableOff returns a boolean indicating whether to force set TTL_ENABLE='OFF' when restoring a TTL table
 func (rf RestoreFlags) HasRestoreWithTTLEnableOff() bool {
 	return rf.has(RestoreWithTTLEnableOff)
+}
+
+
+// HasRestoreForNonPrepPlanCache returns a boolean indicating whether `rf` has `RestoreForNonPrepPlanCache` flag.
+func (rf RestoreFlags) HasRestoreForNonPrepPlanCache() bool {
+	return rf.has(RestoreForNonPrepPlanCache)
 }
 
 // RestoreCtx is `Restore` context to hold flags and writer.