--- a/planner/core/plan_cache_param_test.go
+++ b/planner/core/plan_cache_param_test.go
@@ -16,63 +16,68 @@ package core
 
 import (
 	"context"
-	"strings"
 	"testing"
 
 	"github.com/pingcap/tidb/parser"
-	"github.com/pingcap/tidb/parser/format"
 	"github.com/stretchr/testify/require"
 )
 
 func TestParameterize(t *testing.T) {
 	sctx := MockContext()
 	cases := []struct {
-		sql        string
-		paramSQL   string
-		params     []interface{}
-		restoreSQL string
+		sql      string
+		paramSQL string
+		params   []interface{}
 	}{
 		{
 			"select * from t where a<10",
 			"SELECT * FROM t WHERE a<?",
 			[]interface{}{int64(10)},
-			"SELECT * FROM `t` WHERE `a`<10",
 		},
 		{
 			"select * from t",
 			"SELECT * FROM t",
 			[]interface{}{},
-			"SELECT * FROM `t`",
 		},
 		{
 			"select * from t where a<10 and b<20 and c=30 and d>40",
 			"SELECT * FROM t WHERE a<? AND b<? AND c=? AND d>?",
 			[]interface{}{int64(10), int64(20), int64(30), int64(40)},
-			"SELECT * FROM `t` WHERE `a`<10 AND `b`<20 AND `c`=30 AND `d`>40",
 		},
 		{
 			"select * from t where a='a' and b='bbbbbbbbbbbbbbbbbbbbbbbb'",
 			"SELECT * FROM t WHERE a=? AND b=?",
 			[]interface{}{"a", "bbbbbbbbbbbbbbbbbbbbbbbb"},
-			"SELECT * FROM `t` WHERE `a`=_UTF8MB4'a' AND `b`=_UTF8MB4'bbbbbbbbbbbbbbbbbbbbbbbb'",
 		},
 		{
 			"select 1, 2, 3 from t where a<10",
 			"SELECT 1,2,3 FROM t WHERE a<?",
 			[]interface{}{int64(10)},
-			"SELECT 1,2,3 FROM `t` WHERE `a`<10",
 		},
 		{
 			"select a+1 from t where a<10",
 			"SELECT a+1 FROM t WHERE a<?",
 			[]interface{}{int64(10)},
-			"SELECT `a`+1 FROM `t` WHERE `a`<10",
 		},
 		{
 			"select a+1, sum(b) from t where a<10 group by a+1",
 			"SELECT a+1,sum(b) FROM t WHERE a<? GROUP BY a+1",
 			[]interface{}{int64(10)},
-			"SELECT `a`+1,SUM(`b`) FROM `t` WHERE `a`<10 GROUP BY `a`+1",
+		},
+		{
+			`select a+ "a b c" from t`,
+			`SELECT a+ "a b c" FROM t`,
+			[]interface{}{},
+		},
+		{
+			`select a + 'a b c'+"x" from t`,
+			`SELECT a + 'a b c'+"x" FROM t`, // keep the original format for select fields
+			[]interface{}{},
+		},
+		{
+			`insert into t (a, B, c, D) values (1, 2, 3, 4)`,
+			`INSERT INTO t (a,B,c,D) VALUES (?,?,?,?)`,
+			[]interface{}{int64(1), int64(2), int64(3), int64(4)},
 		},
 		// TODO: more test cases
 	}
@@ -87,12 +92,34 @@ func TestParameterize(t *testing.T) {
 		for i := range params {
 			require.Equal(t, c.params[i], params[i].Datum.GetValue())
 		}
+	}
+}
 
-		err = RestoreASTWithParams(context.Background(), sctx, stmt, params)
-		require.Nil(t, err)
-		var buf strings.Builder
-		rCtx := format.NewRestoreCtx(format.DefaultRestoreFlags, &buf)
-		require.Nil(t, stmt.Restore(rCtx))
-		require.Equal(t, c.restoreSQL, buf.String())
+func BenchmarkParameterizeInsert(b *testing.B) {
+	newOrderInsertOrder := `INSERT INTO orders (o_id, o_d_id, o_w_id, o_c_id, o_entry_d, o_ol_cnt, o_all_local) VALUES (1, 2, 3, 4, 5, 6, 7)`
+	stmt, err := parser.New().ParseOneStmt(newOrderInsertOrder, "", "")
+	require.Nil(b, err)
+	sctx := MockContext()
+	_, _, err = ParameterizeAST(context.Background(), sctx, stmt)
+	require.Nil(b, err)
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		ParameterizeAST(context.Background(), sctx, stmt)
+	}
+}
+
+func BenchmarkParameterizeSelect(b *testing.B) {
+	paymentSelectCustomerForUpdate := `SELECT c_first, c_middle, c_last, c_street_1, c_street_2, c_city, c_state, c_zip, c_phone,
+c_credit, c_credit_lim, c_discount, c_balance, c_since FROM customer WHERE c_w_id = ? AND c_d_id = ?AND c_id = ? FOR UPDATE`
+	stmt, err := parser.New().ParseOneStmt(paymentSelectCustomerForUpdate, "", "")
+	require.Nil(b, err)
+	sctx := MockContext()
+	_, _, err = ParameterizeAST(context.Background(), sctx, stmt)
+	require.Nil(b, err)
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		ParameterizeAST(context.Background(), sctx, stmt)
 	}
 }